+++ src/main/kotlin/dev/pandier/hytadle/HytadleRuntimeConfig.kt
@@ -1,6 +1,7 @@
 package dev.pandier.hytadle
 
 import org.gradle.api.provider.Property
+import org.gradle.api.tasks.SourceSet
 import org.gradle.jvm.toolchain.JavaLauncher
 import org.gradle.process.JavaExecSpec
 
@@ -9,6 +10,7 @@ interface HytadleRuntimeConfig : JavaExecSpec {
     val authMode: Property<String>
     val disableSentry: Property<Boolean>
     val javaLauncher: Property<JavaLauncher>
+    val source: Property<SourceSet>
 
     fun copyTo(target: JavaExecSpec)
 }